Split out Head-Related Structural Operations...
on Expressions
