+++ mathics/core/systemsymbols.py
@@ -230,8 +230,8 @@ SymbolSimplify = Symbol("System`Simplify")
 SymbolSin = Symbol("System`Sin")
 SymbolSinh = Symbol("System`Sinh")
 SymbolSlot = Symbol("System`Slot")
-SymbolSparseArray = Symbol("System`SparseArray")
 SymbolSortBy = Symbol("System`SortBy")
+SymbolSparseArray = Symbol("System`SparseArray")
 SymbolSplit = Symbol("System`Split")
 SymbolSqrt = Symbol("System`Sqrt")
 SymbolSqrtBox = Symbol("System`SqrtBox")
